+++ CHANGELOG.md
@@ -14,6 +14,7 @@ develop
 
 - Update “Julia” and “Swift” syntax styles.
 - Optimize text rendering performance a bit.
+- Apply the change of line height/tab width to all split editors so that split editors not focused also can layout text correctly after the change.
 
 
 ### Fixes